User/requirements.txt
pymongo
fastapi
minio
uvicorn
pydantic_settings
pydantic[email]
pillow==11.0.0
python-dotenv